Add Rosenbluth and PERM MC methods
With wrappings

Rosenbluth is non-recursive, and PERM is.
PERM with weight_thresholds_low = [0] and weight_threshold_high = [inf]
is equivalent to Rosenbluth.

To compute the weight thresholds, Rosenbluth, or the default
weight_thresholds can be used
The partition function for a chain of m monomers can be computed, the
mean over N independent experiments/runs.
<Z_m>_N = sum(weight_n)/N
weight_thresholds_low  = c * <Z_m>
weight_thresholds_high  = C * <Z_m>
where the ratio of the constants should be kept constant at C/c ~= 10

TODO: Introduce an energy function to the weight computation.
Right now, the weights only depends on the atmoshphere of the endchain,
i.e. the number of directions available for the endchain to grow
(without intersecting itself).
Adding a Boltzmann factor to the weight: exp(-kT*energy(chain))
So the weight for a chain of m monomers:
    W_m = W_{m-1} * W_m_atmosphere * W_m_boltzmann
where energy(chain) is the function that can be provided.
if energy(chain) = 0 for all conformations, we get the currently
implemented weights.

energy(chain) can be a function related to Wormlike-Chain model, or
similar.